email-scheduler/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
motor==3.3.1
pydantic>=2.6.4
pydantic-settings>=2.12.0
email-validator>=2.2.0
celery==5.6.2
redis==7.1.0
aiosmtplib==5.0.0
python-multipart>=0.0.9
requests>=2.31.0

